+++ Headerfile/CurrentProtection.h
@@ -23,6 +23,6 @@ void setup_currentMeasurementValues(void);
  * @param adc_value The ADC value representing the current measurement.
  * @return 1 if the current is within the acceptable range, 0 otherwise (indicating a failure).
  */
- float calculate_current(uint32_t adc_value);
+uint32_t calculate_current(uint32_t adc_value);
 uint32_t check_overCurrent(uint32_t current);
 #endif // CURRENT_MEASUREMENT_H
